Unify sitelen cogs
