+++ src/ilo/cogs/sitelen/__init__.py
@@ -1,10 +1,5 @@
-from .preview import CogPreview
-from .sp import CogSp
-from .ss import CogSs
+from .cog import CogSitelen
 
 
 def setup(bot):
-    bot.add_cog(CogPreview(bot))
-    bot.add_cog(CogSp(bot))
-    bot.add_cog(CogSs(bot))
-
+    bot.add_cog(CogSitelen(bot))
